Reorg conftest fixtures for functional tests
