+++ tests/conftest.py
@@ -1,0 +1,20 @@
+import pytest
+from h_matchers import Any
+
+
+@pytest.fixture
+def pyramid_settings():
+    return {
+        "client_embed_url": "http://hypothes.is/embed.js",
+        "nginx_server": "http://via3.hypothes.is",
+        "via_html_url": "https://viahtml3.hypothes.is/proxy",
+        "checkmate_url": "http://localhost:9099",
+    }
+
+
+def assert_cache_control(headers, cache_parts):
+    """Assert that all parts of the Cache-Control header are present."""
+    assert dict(headers) == Any.dict.containing({"Cache-Control": Any.string()})
+    assert (
+        headers["Cache-Control"].split(", ") == Any.list.containing(cache_parts).only()
+    )
